root/backend/requirements.txt
fitz==0.0.1.dev2
pandas==2.2.1
Pillow==10.2.0
PyPDF2==3.0.1
pytesseract==0.3.10
Requests==2.31.0
